+++ release_notes.md
@@ -1,3 +1,8 @@
+### Release Notes 1.5.32
+- Ephys pipeline:
+    - passive extractor
+    - units QC
+
 ### Release Notes 1.5.31 Hotfix
 - add the available space of system and raid volumes in local servers reports.